support of filtered options : KP-7286

+++ src/Core/SharedKernels/DataCollection/DataCollection/Implementation/Aggregates/Interview.cs
@@ -4725,7 +4725,7 @@ namespace WB.Core.SharedKernels.DataCollection.Implementation.Aggregates
             }
         }
 
-        public IEnumerable<CategoricalOption> GetFilteredOptionsForQuestion(Identity question, long? parentQuestionValue, string filter)
+        public IEnumerable<CategoricalOption> GetFilteredOptionsForQuestion(Identity question, int? parentQuestionValue, string filter)
         {
             IQuestionnaire questionnaire = this.GetQuestionnaireOrThrow(questionnaireId, questionnaireVersion);
             var filteredOptions = questionnaire.GetOptionsForQuestion(question.Id, parentQuestionValue, filter);
