+++ src/Core/SharedKernels/DataCollection/DataCollection/Repositories/IQuestionOptionsRepository.cs
@@ -6,6 +6,6 @@ namespace WB.Core.SharedKernels.DataCollection.Repositories
 {
     public interface IQuestionOptionsRepository
     {
-        IEnumerable<CategoricalOption> GetOptionsForQuestion(IQuestionnaire questionnaire, Guid questionId, long? parentQuestionValue, string filter);
+        IEnumerable<CategoricalOption> GetOptionsForQuestion(IQuestionnaire questionnaire, Guid questionId, int? parentQuestionValue, string filter);
     }
 }
